# Conflicts:
#	package-lock.json
#	src/redux/root.saga.js
#	src/redux/rootReducer.js
#	src/routes/routes.js

+++ get-env.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
-rm /usr/share/nginx/html/env.js
+if [ -f /usr/share/nginx/html/env.js ] ; then
+    rm /usr/share/nginx/html/env.js
+fi
 touch /usr/share/nginx/html/env.js
 echo "window.env = {}; window.env.REACT_APP_API_URL = '${REACT_APP_API_URL}'" >> /usr/share/nginx/html/env.js
